Fixed not accounting for utf-8 encodings and a relocated database.
Since the database was no longer in /tmp/ but alongside our files, the database location string has been changed accordingly.

Also, to account for the incorrect utf-8 encodings, a simple '.decode(utf-8)' has been slapped on, everywhere that needed it. It's not a pretty solution, I hate it in fact, but it works.