Added the refactored, more general, test-suite made by Helge.
The tests try to access localhost:5000 by default, so I also modified minitwit.py to run on said url. Another thing to note is, for the tests to work, you need to be running the webserver, and also, the tests make changes to the database, which need to be reverted if you want the tests to run successfully a 2nd time.