Added add_message endpooint
Added add_message endpoint, it takes text from body and appends it to the message when endpoint is queried. A bit doubtful about whether we can get the tests to pass, the features are in practice there but wiring up the endpoints to match the refactored_minitwit_tests.py, will maybe prove to be harder said than done.