Fixed displaying of user profiles
Moved username query to the bottom, as it could intepret other queries as user queries if it gets to them first. Also fixed such that we use the helper query_db instead of @db.execute.  Also fixed such that user profiles can now be displayed, this was due to use forgetting to put '@' in front of "profile_user" thus not using the instance variable from Sinatra.