Adjusted workflow run events
Adjusted workflow run events, such that we do not deploy every time something is pushed to main. But instead we evaluate if it has been pushed with a version tag.

Test workflow now only runs when a pull request to main has been made, but is still able to be run manually if developers desire to test their branch.